fix(core): task-3828 CONSIST-1 money_float 10 -> 15 증가분 5건 해소
check_code_ratchets 적색(ci/48ec858bf68c, culprit 60ba097f)의 실제 원인은
check_consistency.py의 money_float 지표였다 -- LiquidationPolicy.max_slice_notional
(실사용처 liquidation_planner.py가 이미 Decimal(str(...))로 방어적 변환하고
있었다)과 StrategyIntent.qty(주문 의도 수량)는 진짜 금액/수량 필드라 Decimal로
바꿨다. metrics_registry.Counter/Gauge의 `amount` 파라미터는 통화가 아니라
범용 관측 지표 증분값이라 `delta`로 개명해 오탐(money 키워드 매칭)만 제거했다
(Prometheus 텍스트 노출 포맷 자체가 float 산술을 전제해 Decimal 전환은
불가능/불필요).

mandates/contracts/v1.py·domain/models.py의 cash_buffer_pct류 3+2건은
같은 스캔에서 함께 나열되지만(알파벳 정렬 + 기본 --top 10 절단의 우연) 이번
증가분이 아니라 기존 baseline(10) 소속 부채이며, OpenAPI 응답 스키마
필드(check_openapi_compat.py의 MAJOR 대상)라 건드리지 않았다 -- 위 5건만
고쳐 money_float이 15 -> 10(baseline)으로 정확히 복귀함을 확인했다.

회귀 테스트: tests/unit/scripts/test_check_consistency.py에 실제 저장소
3개 파일을 스캔해 재발을 잡는 가드 추가. tests/unit/core/script/
test_builtins_strategy.py는 새 decimal import를 순수성 허용 목록에 추가.

게이트: ruff/mypy(src 전체)/check_zone_manifest/check_consistency.py 전부
로컬 통과(money_float 10, baseline 유지).

+++ src/core/loader/risk_policy_loader.py
@@ -5,8 +5,10 @@ Spec: 07_logging_config_v1.3.md#§7.2
 8.2-B 원칙 — Risk 수치는 코드에 하드코딩하지 않고 이 파일로 관리한다.
 Loader.load_config()(5.1)로 읽은 dict를 여기서 Pydantic 모델로 검증한다.
 """
+
 from __future__ import annotations
 
+from decimal import Decimal
 from pathlib import Path
 from typing import TYPE_CHECKING
 
@@ -131,7 +133,7 @@ class LiquidationPolicy(_StrictModel):
     size_jitter_pct: float = _PCT
     interval_min_sec: float = Field(gt=0)
     interval_max_sec: float = Field(gt=0)
-    max_slice_notional: float = Field(gt=0)
+    max_slice_notional: Decimal = Field(gt=0)
     limit_tolerance_bps: float = Field(gt=0)
     slice_ttl_sec: float = Field(gt=0)
     adverse_move_abort_pct: float = _PCT
